+++ lib/database/lib/model/backup-destination-details.ts
@@ -44,6 +44,21 @@ export interface BackupDestinationDetails {
    */
   "dbrsPolicyId"?: string;
   /**
+   * Indicates if backup retention is locked for all the database backups in the Autonomous Container Database (ACD). The retention window cannot be decreased if the backup retention lock is enabled.
+   * Once applied on the Autonomous Container Database, the retention lock cannot be removed, or the retention period cannot be decreased after a 14-day period.
+   * If the backup is a Long Term Backup and retention lock is enabled, the backup cannot be deleted and must expire.
+   * The retention lock set on the Autonomous Container Database is not applicable for cross region remote backups and backups hosted on recovery Appliance backup destination.
+   *
+   */
+  "isRetentionLockEnabled"?: boolean;
+  /**
+   * Defines the automatic and manual backup retention policy for the Autonomous Database termination.
+   * The retention policy set on the Autonomous Container Database is not applicable for cross region remote backups and backups hosted on recovery Appliance backup destination.
+   * Options are 'RETAIN_PER_RETENTION_WINDOW' or 'RETAIN_FOR_72_HOURS'.The default value is 'RETAIN_FOR_72_HOURS'.
+   *
+   */
+  "backupRetentionPolicyOnTerminate"?: BackupDestinationDetails.BackupRetentionPolicyOnTerminate;
+  /**
    * Indicates whether the backup destination is cross-region or local.
    */
   "isRemote"?: boolean;
@@ -64,6 +79,16 @@ export namespace BackupDestinationDetails {
     Local = "LOCAL",
     Dbrs = "DBRS",
     AwsS3 = "AWS_S3",
+    /**
+     * This value is used if a service returns a value for this enum that is not recognized by this
+     * version of the SDK.
+     */
+    UnknownValue = "UNKNOWN_VALUE"
+  }
+
+  export enum BackupRetentionPolicyOnTerminate {
+    RetainPerRetentionWindow = "RETAIN_PER_RETENTION_WINDOW",
+    RetainFor72Hours = "RETAIN_FOR_72_HOURS",
     /**
      * This value is used if a service returns a value for this enum that is not recognized by this
      * version of the SDK.