+++ lib/database/lib/model/cloud-autonomous-vm-cluster.ts
@@ -224,10 +224,15 @@ Example: {@code {\"Department\": \"Finance\"}}
    */
   "dbNodeStorageSizeInGBs"?: number;
   /**
-   * The amount of memory (in GBs) to be enabled per OCPU or ECPU.
+   * The amount of memory (in GBs rounded off to nearest integer value) enabled per ECPU or OCPU. This is deprecated. Please refer to memoryPerComputeUnitInGBs for accurate value.
    *  Note: Numbers greater than Number.MAX_SAFE_INTEGER will result in rounding issues.
    */
   "memoryPerOracleComputeUnitInGBs"?: number;
+  /**
+   * The amount of memory (in GBs) to be enabled per OCPU or ECPU.
+   *  Note: Numbers greater than Number.MAX_SAFE_INTEGER will result in rounding issues.
+   */
+  "memoryPerComputeUnitInGBs"?: number;
   /**
    * The list of [OCIDs](https://docs.oracle.com/iaas/Content/General/Concepts/identifiers.htm) of the Db servers.
    */